requirements.txt
scikit-learn
scipy
PyYAML
opencv-python
configparser
matplotlib
ConfigArgParse
tqdm
tensorflow-gpu==1.13.2
numpy==1.16.1
torch>=1.2
smplx==0.1.13